move SharedSegments to Polyline and include optional flag for closed shapes
