+++ Elements/src/Geometry/Polyline.cs
@@ -501,6 +501,59 @@ namespace Elements.Geometry
             }
             return true;
         }
+
+        /// <summary>
+        /// Identify any shared segments between two polylines.
+        /// </summary>
+        /// <param name="a">The first polyline to compare.</param>
+        /// <param name="b">The second polyline to compare.</param>
+        /// <param name="isClosed">Flag as closed to include segment between first and last vertex.</param>
+        /// <returns>Returns a list of tuples of indices for the segments that match in each polyline.</returns>
+        public static List<(int indexOnA, int indexOnB)> SharedSegments(Polyline a, Polyline b, bool isClosed = false)
+        {
+            var result = new List<(int, int)>();
+
+            // Abbreviate lists to compare
+            var va = a.Vertices;
+            var vb = b.Vertices;
+
+            for (var i = 0; i < va.Count; i++)
+            {
+                var ia = va[i];
+                var ib = va[(i + 1) % va.Count];
+
+                var iterations = isClosed ? vb.Count : vb.Count - 1;
+
+                for (var j = 0; j < iterations; j++)
+                {
+                    var ja = vb[j];
+
+                    if (ia.IsAlmostEqualTo(ja))
+                    {
+                        // Current vertices match, compare next vertices
+                        var jNext = (j + 1) % vb.Count;
+                        var jPrev = j == 0 ? vb.Count - 1 : j - 1;
+
+                        var jb = vb[jNext];
+                        var jc = vb[jPrev];
+
+                        if (ib.IsAlmostEqualTo(jb))
+                        {
+                            // Match is current segment a and current segment b
+                            result.Add((i, j));
+                        }
+
+                        if (ib.IsAlmostEqualTo(jc))
+                        {
+                            // Match is current segment a and previous segment b
+                            result.Add((i, jPrev));
+                        }
+                    }
+                }
+            }
+
+            return result;
+        }
     }
 
     /// <summary>
@@ -530,7 +583,7 @@ namespace Elements.Geometry
         /// </summary>
         /// <param name="l">The line to convert.</param>
         public static Polyline ToPolyline(this Line l) => new Polyline(new[] { l.Start, l.End });
-   
+
     }
 
     /// <summary>
@@ -539,7 +592,7 @@ namespace Elements.Geometry
     public enum EndType
     {
         /// <summary>
-        /// Open ends are extended by the offset distance and squared off 
+        /// Open ends are extended by the offset distance and squared off
         /// </summary>
         Square,
         /// <summary>
